made eval more robust
