+++ frontend/streamlit.py
@@ -2,6 +2,7 @@ import os
 import time
 import streamlit as st
 from .pdf import serve_pdf
+from .questions import check_baseline_answerable
 from backend.inference import chat_completion
 from backend.statistics import (
     init_user_session,
@@ -11,47 +12,9 @@ from backend.statistics import (
 from .utils import (
     load_css,
     handle_feedback,
-    display_confusion_matrix,
+    get_feedback_question,
+    display_confusion_matrix
 )
-
-# Define constants for baseline questions
-BASELINE_QUESTIONS = {
-    # 10 Answerable questions
-    "Who is Hironori Washizaki?": True,
-    "How does software testing impact the overall software development lifecycle?": True,
-    "What is the agile methodology?": True,
-    "What are the different types of software models, and when should each be used?": True,
-    "How does software configuration management ensure project success?": True,
-    "What role do user requirements play in software design and architecture?": True,
-    "What is software dependability?": True,
-    "How does project management in software engineering differ from traditional project management?": True,
-    "What strategies can be used for effective risk management in software engineering projects?": True,
-    "What is the purpose of static analysis in software testing?": True,
-
-    # 10 Unanswerable Questions
-    "What is the largest possible number that could exist?": False,
-    "How will AI evolve in the next 100 years?": False,
-    "What is the solution to the Riemann Hypothesis?": False,
-    "What is the smallest possible Turing machine?": False,
-    "What is the most complex algorithm that can never be solved?": False,
-    "What is the most efficient way to handle an infinite stream of data?": False,
-    "Is there a way to build a fully self-sustaining human colony on Mars with current technology?": False,
-    "What's the upper limit of computational power for classical computers?": False,
-    "How could we fully eliminate all types of noise in wireless communications?": False,
-    "How can we create a material that is completely indestructible?": False
-
-}
-
-def get_feedback_question(prompt: str) -> str:
-    """Return feedback question based on baseline questions."""
-    if prompt in BASELINE_QUESTIONS:
-        is_answerable = BASELINE_QUESTIONS[prompt]
-        return (
-            "Did the chatbot correctly answer this answerable question?"
-            if is_answerable
-            else "Did the chatbot correctly answer this unanswerable question?"
-        )
-    return "Was this response helpful?"
 
 def initialize_session():
     """Initialize user session and handle errors."""
@@ -130,6 +93,7 @@ def generate_response(prompt: str, response_container):
 
 def save_conversation_to_db(prompt: str, response: str) -> int:
     """Save conversation to the database and return conversation ID."""
+    answerable = check_baseline_answerable(prompt)
     return insert_conversation(
         question=prompt,
         response=response,
@@ -139,7 +103,7 @@ def save_conversation_to_db(prompt: str, response: str) -> int:
         response_time=st.session_state.response_time,
         correct=None,
         user_id=st.session_state.user_id,
-        answerable=BASELINE_QUESTIONS.get(prompt, None),
+        answerable=answerable
     )
 
 def update_session_messages(prompt: str, response: str, conversation_id: int):
